refactor(desktop): keep the codex_message_items hydration fallback; drop the placeholder half
- The placeholder ("assistant response not persisted") is removed: it turned a
  wholly-empty persisted row into TEXT, which in reconcileResumeMessages could
  replace a text-only live stream row at the same ordinal until the turn
  settled. The sidecar fallback alone covers the #68321 field repro.
- Phase filter matches the backend (codex_responses_adapter _OutputScan._message):
  commentary AND analysis are reasoning-channel narration, not the reply.
- Tests trimmed to the two contracts that are red on main; six that passed on
  main (pre-existing reasoning-part behaviour) are gone. File renamed for what
  it now covers.
